requirements.txt
# AI Image Caption Generator — Python Dependencies

# Core AI / ML
transformers>=4.38.0
torch>=2.2.0
torchvision>=0.17.0

# Image handling
Pillow>=10.0.0

# HTTP
requests>=2.31.0

# Web UI (optional — for app.py)
gradio>=4.20.0

# REST API (optional — for api.py)
flask>=3.0.0
